requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AI/ML
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
openai==1.12.0

# Google APIs
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Scheduling
apscheduler==3.10.4

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
python-dateutil==2.8.2

# Logging
structlog==24.1.0

# Google Drive backup
google-auth==2.27.0
